fix missing received values when checking option
For example, an error would say:
AssertionError: Invalid server_id[54] option received:  but expected 10.0.122.32
The empty received value does not mean that the option is not there. The
error would have been different if that was the case. There is
a received value, but forge does not know how to display it.

This change turns the error into:
AssertionError: Invalid server_id[54] option received: ('server_id', '10.0.122.70') but expected 10.0.122.32