use in-subnet reservations in test_reservation_mode_in_globals
According to https://kea.readthedocs.io/en/kea-2.3.5/arm/dhcp4-srv.html#global-reservations-in-dhcpv4:

An address assigned via global host reservation must be feasible for the
subnet the server selects for the client. In other words, the address must
lie within the subnet otherwise it will be ignored and the server will
attempt to dynamically allocate an address.  In the event the selected subnet
belongs to a shared-network the server will check for feasibility against
the subnet's siblings, selecting the first in-range subnet.  If no such
subnet exists, the server will fallback to dynamically allocating the address.