adapt test to order of radius attributes
Multiple instances of certain attributes such as Framed-Pool are
uncharted-territory/undefined-behavior according to RFC which says that
it SHOULD contain a maximum of 1. As a result, Kea has some leeway in
the implementation, and, as it turns out the new RADIUS hook library
only considers the last attribute, while the old RADIUS hook library
only considered the first attribute. This is documented here:
https://kea.readthedocs.io/en/kea-2.5.4/arm/hooks.html#differences-between-radius-hook-libraries-prior-to-2-4-0-and-as-of-2-6-0