[#463] Ongoing Maintenance: Leave only the custom entries in pylint.rc
pylint.rc was created with "pylint --generate-rcfile". It has a lot of
explicit default (if that's not an oxymoron) entries. It's difficult to
know what we set explicitly and what is default. Also, when a default
changes in pylint, maybe it was for good reason, and maybe it would be good to let it affect us, instead of silently keeping the old default.